This commit adds formatting and links to log output
When running pharos with many older, unresponded-to threads, it's hard
to find the threads that are triggering the notifications. This PR adds
a permalink to the convo to the log output so you can go back and find
them.

I also cleaned up the Dockerfile a bit and extended the .gitignore.